refactor: move BackupRange and CheckpointEntry to api package
