+++ zeebe/backup/src/main/java/io/camunda/zeebe/backup/api/Checkpoint.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.zeebe.backup.api;
+
+import io.camunda.zeebe.protocol.record.value.management.CheckpointType;
+
+public record Checkpoint(
+    CheckpointType checkpointType,
+    long checkpointId,
+    long checkpointTimestamp,
+    long checkpointPosition,
+    long firstLogPosition) {}
